Break up graphs into multiple files
